backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.3",
    "django-filter>=25.1",
    "django-rest-framework>=0.1.0",
    "django-stubs>=5.2.1",
]
